% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarrMot-package.R
\docType{package}
\name{MarrMot-package}
\alias{MarrMot}
\alias{MarrMot-package}
\title{MarrMot: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
