% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarrMot_calib.R
\name{MarrMot_calib}
\alias{MarrMot_calib}
\title{MarrMot_calib}
\usage{
MarrMot_calib(df, models, maxn = 10)
}
\arguments{
\item{df}{A data.table with the columns of c("prcp", "pet", "tair").}

\item{models}{model names}
}
\description{
MarrMot_calib
}
