% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarrMot_calib.R
\name{MarrMot_predict}
\alias{MarrMot_predict}
\title{MarrMot_predict}
\usage{
MarrMot_predict(df, par, model)
}
\arguments{
\item{df}{A data.table with the columns of c("prcp", "pet", "tair").}

\item{par}{numberic vector or scalar}

\item{model}{model name}
}
\description{
MarrMot_predict
}
